fix test and move writer to gen folder
