+++ transfuse-core/src/main/java/org/androidtransfuse/gen/WriterOutputStream.java
@@ -1,4 +1,4 @@
-package org.androidtransfuse.util.apache.commons;
+package org.androidtransfuse.gen;
 
 
 import java.io.IOException;
