+++ transfuse/src/test/java/org/androidtransfuse/gen/FilerResourceWriterTest.java
@@ -16,7 +16,6 @@
 package org.androidtransfuse.gen;
 
 import com.sun.codemodel.JCodeModel;
-import org.apache.commons.io.output.WriterOutputStream;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -25,6 +24,7 @@ import javax.tools.FileObject;
 import javax.tools.StandardLocation;
 import java.io.IOException;
 import java.io.OutputStream;
+import java.io.Writer;
 
 import static org.junit.Assert.assertEquals;
 import static org.mockito.Mockito.*;
@@ -32,6 +32,7 @@ import static org.mockito.Mockito.*;
 /**
  * @author John Ericksen
  */
+
 public class FilerResourceWriterTest {
 
     private static final String TEST_PACKAGE = "org.test";
@@ -41,6 +42,7 @@ public class FilerResourceWriterTest {
     private Filer mockFiler;
     private FileObject mockFile;
     private OutputStream mockOutputStream;
+    private Writer mockWriter;
     private JCodeModel codeModel;
 
     @Before
@@ -48,16 +50,19 @@ public class FilerResourceWriterTest {
         mockFiler = mock(Filer.class);
         mockFile = mock(FileObject.class);
         mockOutputStream = mock(OutputStream.class);
+        mockWriter = mock(Writer.class);
 
-        resourceWriter = new FilerResourceWriter(mockFiler);
+        resourceWriter = spy(new FilerResourceWriter(mockFiler));
         codeModel = new JCodeModel();
     }
 
     @Test
     public void testCreateResource() throws IOException {
-
         when(mockFiler.createResource(StandardLocation.SOURCE_OUTPUT, TEST_PACKAGE, TEST_FILENAME)).thenReturn(mockFile);
-        when(mockFile.openOutputStream()).thenReturn(mockOutputStream);
-
+        doReturn(mockWriter).when(mockFile).openWriter();
+        doReturn(mockOutputStream).when(resourceWriter).getWriterOutputStream(mockWriter);
+        assertEquals(mockOutputStream, resourceWriter.openBinary(codeModel._package(TEST_PACKAGE), TEST_FILENAME));
+        resourceWriter.close();
+        verify(mockOutputStream).close();
     }
 }
